feat(quiz): implement ResultModal component for displaying quiz results with animations
